requirements.txt
flask
flask_sqlalchemy
flask_login
mysql-connector-python
werkzeug
